[flink] Provide scan.push-down to disable source pushdown (#1516)
